User update/delete

+++ service/user_delete.py
@@ -6,6 +6,10 @@ def user_delete(collection, request):
     try:
         myquery = { "_id": ObjectId(request["id"])}
 
+        cursor = collection.find(myquery)
+        list_cur = list(cursor)
+        if list_cur == []:
+            raise Exception("User does not exist")
         collection.delete_one(myquery)
         
         return validation_response(True, "Success Delete User", 200)
